Controller/requirements.txt
Pillow==7.0.0
pygame==2.0.1
